refactor(PurchaseAmount): purchaseAmount 값 int -> long 형변환
- 해당 과제는 int 이상의 값이 들어올 경우 NumberParser 에서 예외처리
- 그러나 확장성을 고려하여 2147483647원 이상의 값이 입력될 수 있음을 고려하여 수정

+++ src/main/java/lotto/domain/PurchaseAmount.java
@@ -5,7 +5,7 @@ import lotto.domain.exception.PurchaseNotAvailableException;
 
 public class PurchaseAmount {
     public static final int LOTTO_PRICE = 1000;
-    private final int purchaseAmount;
+    private final long purchaseAmount;
 
     private PurchaseAmount(final int purchaseAmount) {
         validateInputValue(purchaseAmount);
@@ -41,11 +41,11 @@ public class PurchaseAmount {
         return purchaseAmount % LOTTO_PRICE != 0;
     }
 
-    public int getPurchaseAmount() {
+    public long getPurchaseAmount() {
         return purchaseAmount;
     }
 
     public int getPurchaseQuantity() {
-        return purchaseAmount / LOTTO_PRICE;
+        return (int) purchaseAmount / LOTTO_PRICE;
     }
 }
